+++ app/composables/ving/useAdminLinks.mjs
@@ -8,6 +8,7 @@
 export default () => {
     const links = computed(() => {
         const out = [
+            { label: 'Cron Jobs', to: '/cronjob', icon: 'ph:clock' },
             { label: 'System Wide Alert', to: '/system-wide-alert', icon: 'ph:megaphone' },
             { label: 'Users', to: '/user/admin', icon: 'ph:users' },
         ];