+++ ving/docs/change-log.md
@@ -8,6 +8,9 @@ outline: deep
 ### 2024-07-04
 * Added `options` param to ving schema props.
 * Fixed generating pathing for pages.
+* Added Cron Jobs subsystem.
+* NOTE: Run a database migration to add the new cronjob table.
+* Fixed: Can't have duplicate cron specs in redis. #175
 
 ### 2024-07-01
 * Added verifiedEmail field to the User CLI.